bad predicate tests removed
No way to provide compile-time checking whether
the custom compare operation guarantee
a strict weak ordering on the values.

25.4 Sorting and related operations [alg.sorting]
...
For algorithms other than those described in 25.4.3 to work
correctly, comp has to induce a strict weak ordering on the values.

The term strict refers to the requirement of an irreflexive
relation (!comp(x, x) for all x), and the term weak
to requirements that are not as strong as those
for a total ordering, but stronger than those
for a partial ordering. If we define equiv(a, b) as
!comp(a, b) && !comp(b, a), then the requirements
are that comp and equiv both be transitive relations:

  - comp(a, b) && comp(b, c) implies comp(a, c)

  - equiv(a, b) && equiv(b, c) implies equiv(a, c)
    [ Note: Under these conditions, it can be shown
    that
    - equiv is an equivalence relation
    - comp induces a well-defined relation on
      the equivalence classes determined by equiv
    - The induced relation is a strict total ordering. -- end note ]